requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
Django==4.2.10
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
setuptools==68.2.2
sniffio==1.3.1
sqlparse==0.4.4
tzdata==2024.1
wheel==0.41.2
